move mock store for Navbar to separate file.
